# backend/swagger-ui/swagger.yaml
swagger: "2.0"
info:
  description: "This is a API doc of udo-jieshixin"
  version: "1.0.0"
  title: "udo-jieshixin"
  contact:
    email: "nemoworks@gmail.com"
host: "localhost:8080"
basePath: "/api"
tags:
- name: "graph"
  description: "Everything about graph engine"
  
schemes:
- "https"
- "http"
paths:
  /graph:
    post:
      tags:
      - "graph"
      summary: "Create a graph by type"
      operationId: "createGraph"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        schema:
          $ref: "#/definitions/Graph"
      responses:
        "405":
          description: "Invalid input"
    get:
      tags:
      - "graph"
      summary: "Get graph list"
      operationId: "getGraphList"
      produces:
      - "application/json"
      responses:
        "405":
          description: "Invalid input"

  /graph/{id}:
    get:
      tags:
      - "graph"
      summary: "Get graph by id"
      operationId: "getGraph"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The name that needs to be fetched. Use user1 for testing. "
        required: true
        type: "string"
      responses:
        "405":
          description: "Invalid input"
    put:
      tags:
      - "graph"
      summary: "Update graph by id"
      operationId: "updateGraph"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The name that needs to be fetched. Use user1 for testing. "
        required: true
        type: "string"
      responses:
        "405":
          description: "Invalid input"
    delete:
      tags:
      - "graph"
      summary: "Delete graph by id"
      operationId: "deleteGraph"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The name that needs to be fetched. Use user1 for testing. "
        required: true
        type: "string"
      responses:
        "405":
          description: "Invalid input"
  
definitions:
  Graph:
    type: "object"
    properties:
      nodes:
        type: "array"
        items: 
          $ref: "#/definitions/Node"
      edges:
        type: "array"
        items:
          $ref: "#/definitions/Edge"
  Node:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
        
  Edge:
    type: "object"
    properties:
      from:
        type: "string"
      to:
        type: "string"

externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"